+++ src/BDSComponentFactory.cc
@@ -44,6 +44,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "BDSSamplerPlane.hh"
 #include "BDSScreen.hh"
 #include "BDSShield.hh"
+#include "BDSTarget.hh"
 #include "BDSTeleporter.hh"
 #include "BDSTerminator.hh"
 #include "BDSTiltOffset.hh"
@@ -328,6 +329,8 @@ BDSAcceleratorComponent* BDSComponentFactory::CreateComponent(Element const* ele
       {component = CreateEllipticalCollimator(); break;} 
     case ElementType::_RCOL:
       {component = CreateRectangularCollimator(); break;}
+    case ElementType::_TARGET:
+      {component = CreateTarget(); break;}
     case ElementType::_JCOL:
       {component = CreateJawCollimator(); break;}
     case ElementType::_MUONSPOILER:    
@@ -1357,6 +1360,22 @@ BDSAcceleratorComponent* BDSComponentFactory::CreateRectangularCollimator()
 				      element->ysizeOut*CLHEP::m,
 				      PrepareColour(element),
 				      circularOuter);
+}
+
+BDSAcceleratorComponent* BDSComponentFactory::CreateTarget()
+{
+  if (!HasSufficientMinimumLength(element))
+    {return nullptr;}
+  G4bool circularOuter = false;
+  G4String apertureType = G4String(element->apertureType);
+  if (apertureType == "circular")
+    {circularOuter = true;}
+  return new BDSTarget(elementName,
+		       element->l*CLHEP::m,
+		       PrepareHorizontalWidth(element),
+		       PrepareMaterial(element),
+		       PrepareColour(element),
+		       circularOuter);
 }
 
 BDSAcceleratorComponent* BDSComponentFactory::CreateEllipticalCollimator()