+++ manual/source/version_history.rst
@@ -65,6 +65,8 @@ New Features
 
 * A new `ct` keyword has been implemented to allow the conversion of DICOM CT images into
   voxelized geometries.
+* New `target` beam line component. We could always create a block of material with a closed
+  `rcol` but this is more intuitive.
 
 **Fields**